refactor(modules): split schemas.py into data_entries/factors/handlers (#1556)
Each Calculator module package now separates entry validation schemas
(data_entries.py), factor DTOs + handlers (factors.py), and
BaseModuleHandler subclasses (handlers.py), with package __init__
re-exports keeping registration side effects. Prepares reuse of the
factor/emission side by the Simulator Plan (#1556) without dragging in
Calculator entry validation. research_facilities' two schema files
merge into the same layout; registration-only seed imports collapse to
package imports.
