+++ backend/app/api/v1/building_rooms.py
@@ -8,7 +8,7 @@ from sqlmodel.ext.asyncio.session import AsyncSession
 from app.api.deps import get_db
 from app.core.security import get_current_active_user
 from app.models.user import User
-from app.modules.buildings.schemas import (
+from app.modules.buildings import (
     BuildingRoomBuildingResponse,
     BuildingRoomResponse,
 )
