+++ backend/app/modules/equipment/handlers.py
@@ -1,131 +1,23 @@
-from typing import Optional, Self
+from typing import Optional
 
-from pydantic import ValidationInfo, field_validator, model_validator
 from sqlalchemy import func
 
 from app.core.config import get_settings
-from app.core.logging import get_logger
 from app.models.data_entry import DataEntry, DataEntryTypeEnum
 from app.models.data_entry_emission import DataEntryEmission, EmissionComputation
 from app.models.factor import Factor
 from app.models.module_type import ModuleTypeEnum
+from app.modules.equipment.data_entries import (
+    MAX_WEEKLY_USAGE_HOURS,
+    EquipmentHandlerCreate,
+    EquipmentHandlerResponse,
+    EquipmentHandlerUpdate,
+)
 from app.schemas.data_entry import (
     BaseModuleHandler,
     DataEntryCreate,
-    DataEntryResponseGen,
     DataEntryUpdate,
 )
-from app.schemas.factor import (
-    BaseFactorHandler,
-    EmissionType,
-    FactorCreate,
-    FactorResponseGen,
-    FactorUpdate,
-)
-
-logger = get_logger(__name__)
-
-MAX_WEEKLY_USAGE_HOURS = 168
-
-
-def _validate_weekly_usage_hours(v: Optional[int]) -> Optional[int]:
-    if v is None:
-        return v
-    if v < 0:
-        raise ValueError("Usage hours must be non-negative")
-    if v > MAX_WEEKLY_USAGE_HOURS:
-        raise ValueError(
-            f"Usage hours cannot exceed {MAX_WEEKLY_USAGE_HOURS} hours per week"
-        )
-    return v
-
-
-def _validate_non_negative_float(
-    v: Optional[float], field_name: str
-) -> Optional[float]:
-    if v is None:
-        return v
-    if v < 0:
-        raise ValueError(f"{field_name} must be non-negative")
-    return v
-
-
-class _EquipmentUsageHoursValidationMixin:
-    active_usage_hours_per_week: Optional[int]
-    standby_usage_hours_per_week: Optional[int]
-
-    @field_validator(
-        "active_usage_hours_per_week", "standby_usage_hours_per_week", mode="after"
-    )
-    @classmethod
-    def validate_usage_hours(cls, v: Optional[int]) -> Optional[int]:
-        return _validate_weekly_usage_hours(v)
-
-    @model_validator(mode="after")
-    def validate_total_usage_hours(self) -> Self:
-        active_hours = self.active_usage_hours_per_week
-        standby_hours = self.standby_usage_hours_per_week
-        if active_hours is not None and standby_hours is not None:
-            if active_hours + standby_hours > MAX_WEEKLY_USAGE_HOURS:
-                raise ValueError(
-                    "The sum of active_usage_hours_per_week and "
-                    "standby_usage_hours_per_week must be <= 168"
-                )
-        return self
-
-
-class _EquipmentFactorValidationMixin(_EquipmentUsageHoursValidationMixin):
-    @field_validator(
-        "active_power_w", "standby_power_w", "ef_kg_co2eq_per_kwh", mode="after"
-    )
-    @classmethod
-    def validate_factor_non_negative(
-        cls, v: Optional[float], info: ValidationInfo
-    ) -> Optional[float]:
-        return _validate_non_negative_float(v, info.field_name or "")
-
-
-## DATA-ENTRY-SPECIFIC SCHEMAS AND HANDLER FOR EQUIPMENT ELECTRIC CONSUMPTION MODULE
-
-
-# https://epfl-enac.github.io/co2-calculator-back-office-doc/data-description/#equipment
-class EquipmentHandlerResponse(DataEntryResponseGen):
-    name: str
-    equipment_class: str
-    sub_class: Optional[str] = None
-    active_usage_hours_per_week: Optional[int] = None
-    standby_usage_hours_per_week: Optional[int] = None
-    note: Optional[str] = None
-    kg_co2eq: Optional[float] = None
-    active_power_w: Optional[int] = None
-    standby_power_w: Optional[int] = None
-
-
-class EquipmentHandlerCreate(_EquipmentUsageHoursValidationMixin, DataEntryCreate):
-    equipment_id: str
-    name: str
-    equipment_class: str
-    sub_class: Optional[str] = None
-    active_usage_hours_per_week: Optional[int] = None
-    standby_usage_hours_per_week: Optional[int] = None
-    note: Optional[str] = None
-    # kg_co2eq: Optional[float] = None  # from csv is __kg_co2eq_override__
-
-    @field_validator("equipment_id", "name", "equipment_class", mode="after")
-    @classmethod
-    def _non_empty(cls, v: str, info: ValidationInfo) -> str:
-        if not v.strip():
-            raise ValueError(f"{info.field_name} cannot be empty")
-        return v.strip()
-
-
-class EquipmentHandlerUpdate(_EquipmentUsageHoursValidationMixin, DataEntryUpdate):
-    active_usage_hours_per_week: Optional[int] = None
-    standby_usage_hours_per_week: Optional[int] = None
-    name: Optional[str] = None
-    equipment_class: Optional[str] = None
-    sub_class: Optional[str] = None
-    note: Optional[str] = None
 
 
 class EquipmentModuleHandler(BaseModuleHandler):
@@ -275,69 +167,3 @@ class EquipmentModuleHandler(BaseModuleHandler):
 
     def validate_update(self, payload: dict) -> DataEntryUpdate:
         return self.update_dto.model_validate(payload)
-
-
-## FACTORS for equipment electric consumption
-
-
-classification_fields: list[str] = ["equipment_class", "sub_class"]
-value_fields: list[str] = [
-    "active_power_w",
-    "standby_power_w",
-    "active_usage_hours_per_week",
-    "standby_usage_hours_per_week",
-    "ef_kg_co2eq_per_kwh",
-]
-
-
-class EquipmentFactorCreate(_EquipmentFactorValidationMixin, FactorCreate):
-    equipment_class: str
-    sub_class: Optional[str] = None
-    active_usage_hours_per_week: int  # make it mandatory
-    standby_usage_hours_per_week: int  # make it mandatory
-    active_power_w: float
-    standby_power_w: float
-    ef_kg_co2eq_per_kwh: float
-    # equipment_category: str  # only for upload Mandatory (checked in csv upload)
-    # equipment_category is the routing column (picks scientific/it/other).
-    # It is consumed in the factor CSV provider, not carried on this DTO —
-    # its presence + case-sensitive {scientific,it,other} enum is enforced
-    # there (see base_factor_csv_provider._resolve_data_entry_type).
-
-
-class EquipmentFactorUpdate(_EquipmentFactorValidationMixin, FactorUpdate):
-    equipment_class: Optional[str] = None
-    sub_class: Optional[str] = None
-    active_power_w: Optional[float] = None
-    standby_power_w: Optional[float] = None
-    active_usage_hours_per_week: Optional[int] = None
-    standby_usage_hours_per_week: Optional[int] = None
-    ef_kg_co2eq_per_kwh: Optional[float] = None
-
-
-class EquipmentFactorResponse(FactorResponseGen):
-    equipment_class: str
-    sub_class: Optional[str] = None
-    active_power_w: float
-    standby_power_w: float
-
-
-class EquipmentFactorHandler(BaseFactorHandler):
-    data_entry_type: DataEntryTypeEnum | None = None
-    registration_keys = [
-        DataEntryTypeEnum.scientific,
-        DataEntryTypeEnum.it,
-        DataEntryTypeEnum.other,
-    ]
-    category_field: str = "equipment_category"
-    emission_type: EmissionType = EmissionType.equipment
-
-    create_dto = EquipmentFactorCreate
-    update_dto = EquipmentFactorUpdate
-    response_dto = EquipmentFactorResponse
-
-    classification_fields: list[str] = classification_fields
-    value_fields: list[str] = value_fields
-
-    def to_response(self, factor: Factor) -> FactorResponseGen:
-        return self.response_dto.model_validate(factor.model_dump)
