+++ backend/app/modules/headcount/__init__.py
@@ -1,1 +1,45 @@
-"""Headcount module package."""
+"""Headcount module package.
+
+Importing this package registers the module and factor handlers
+(metaclass side effect in ``handlers``/``factors``).
+"""
+
+from app.modules.headcount.data_entries import (
+    SIUS_CODE_VALUES,
+    HeadCountCreate,
+    HeadcountItemResponse,
+    HeadcountMemberDropdownItem,
+    HeadCountStudentCreate,
+    HeadCountStudentResponse,
+    HeadCountStudentUpdate,
+    HeadCountUpdate,
+)
+from app.modules.headcount.factors import (
+    HeadcountFactorCreate,
+    HeadcountFactorResponse,
+    HeadcountFactorUpdate,
+    HeadcountMemberFactorHandler,
+    HeadcountStudentFactorHandler,
+)
+from app.modules.headcount.handlers import (
+    HeadcountMemberModuleHandler,
+    HeadcountStudentModuleHandler,
+)
+
+__all__ = [
+    "SIUS_CODE_VALUES",
+    "HeadCountCreate",
+    "HeadCountStudentCreate",
+    "HeadCountStudentResponse",
+    "HeadCountStudentUpdate",
+    "HeadCountUpdate",
+    "HeadcountItemResponse",
+    "HeadcountMemberDropdownItem",
+    "HeadcountFactorCreate",
+    "HeadcountFactorResponse",
+    "HeadcountFactorUpdate",
+    "HeadcountMemberFactorHandler",
+    "HeadcountStudentFactorHandler",
+    "HeadcountMemberModuleHandler",
+    "HeadcountStudentModuleHandler",
+]
