--- backend/app/modules/purchase/schemas.py
@@ -1,568 +1,0 @@
-from typing import Any, Optional
-
-from pydantic import field_validator, model_validator
-
-from app.core.logging import get_logger
-from app.models.data_entry import DataEntry, DataEntryTypeEnum
-from app.models.data_entry_emission import (
-    DataEntryEmission,
-    EmissionComputation,
-)
-from app.models.module_type import ModuleTypeEnum
-from app.modules.emissions import EmissionType
-from app.schemas.data_entry import (
-    BaseModuleHandler,
-    DataEntryCreate,
-    DataEntryResponseGen,
-    DataEntryUpdate,
-)
-from app.schemas.factor import (
-    BaseFactorHandler,
-    FactorCreate,
-    FactorResponseGen,
-    FactorUpdate,
-)
-from app.services.exchange_rates_service import ExchangeRatesService
-
-logger = get_logger(__name__)
-
-
-class PurchaseHandlerResponse(DataEntryResponseGen):
-    name: str
-    supplier: Optional[str] = None
-    quantity: Optional[float] = None
-    total_spent_amount: float
-    currency: Optional[str] = None
-    purchase_institutional_code: Optional[str] = None
-    note: Optional[str] = None
-    kg_co2eq: Optional[float] = None
-
-
-class PurchaseCentralizedHandlerResponse(DataEntryResponseGen):
-    name: str
-    unit: Optional[str] = None
-    annual_consumption: float
-    coef_to_kg: float
-    note: Optional[str] = None
-    kg_co2eq: Optional[float] = None
-
-
-class PurchaseHandlerCreate(DataEntryCreate):
-    name: str
-    supplier: Optional[str] = None
-    quantity: Optional[float] = None
-    total_spent_amount: float
-    currency: Optional[str] = None  # doc say mandatory, but with default -> optional
-    purchase_institutional_code: str
-    purchase_institutional_description: Optional[str] = None
-    purchase_additional_code: Optional[str] = None
-    note: Optional[str] = None
-    # __kg_co2eq_override__ is used to override the kg_co2eq calculation
-
-    @model_validator(mode="before")
-    @classmethod
-    def ensure_default_currency(cls, data: Any) -> Any:
-        """Ensure default currency is applied when input has null or empty currency."""
-        if isinstance(data, dict):
-            currency = data.get("currency")
-            # Apply default when currency is None, empty string, or whitespace-only
-            if currency is None or (
-                isinstance(currency, str) and currency.strip() == ""
-            ):
-                data["currency"] = "chf"
-        return data
-
-    @field_validator("quantity", mode="after")
-    @classmethod
-    def validate_quantity(cls, v: Optional[float]) -> Optional[float]:
-        if v is None:
-            return v
-        if v < 0:
-            raise ValueError("Quantity must be non-negative")
-        return v
-
-    @field_validator("total_spent_amount", mode="after")
-    @classmethod
-    def validate_total_spent_amount(cls, v: float) -> float:
-        if v < 0:
-            raise ValueError("Total spend amount must be non-negative")
-        return v
-
-    @field_validator("purchase_institutional_code", mode="after")
-    @classmethod
-    def validate_purchase_institutional_code(cls, v: str) -> str:
-        if len(v) < 1:
-            raise ValueError(
-                "Purchase institutional code must be at least 1 character long"
-            )
-        return v
-
-    @field_validator("currency", mode="after")
-    @classmethod
-    def validate_currency(cls, v: Optional[str]) -> str:
-        if v is None:
-            return "chf"
-        normalized_v = v.strip().lower()
-        valid_currencies = [
-            "aud",
-            "cad",
-            "chf",
-            "cny",
-            "eur",
-            "gbp",
-            "jpy",
-            "sek",
-            "usd",
-        ]
-        if normalized_v not in valid_currencies:
-            raise ValueError(f"Currency must be one of: {valid_currencies}")
-        return normalized_v
-
-
-class PurchaseCentralizedHandlerCreate(DataEntryCreate):
-    name: str
-    unit: str
-    annual_consumption: float
-    coef_to_kg: float
-    note: Optional[str] = None
-
-    @field_validator("annual_consumption", "coef_to_kg", mode="after")
-    @classmethod
-    def validate_positive(cls, v: float) -> float:
-        if v < 0:
-            raise ValueError("Value must be non-negative")
-        return v
-
-
-class PurchaseHandlerUpdate(DataEntryUpdate):
-    name: Optional[str] = None
-    supplier: Optional[str] = None
-    quantity: Optional[float] = None
-    total_spent_amount: Optional[float] = None
-    currency: Optional[str] = None
-    purchase_institutional_code: Optional[str] = None
-    purchase_additional_code: Optional[str] = None
-    note: Optional[str] = None
-
-    @field_validator("quantity", mode="after")
-    @classmethod
-    def validate_quantity(cls, v: Optional[float]) -> Optional[float]:
-        if v is None:
-            return v
-        if v < 0:
-            raise ValueError("Quantity must be non-negative")
-        return v
-
-    @field_validator("total_spent_amount", mode="after")
-    @classmethod
-    def validate_total_spent_amount(cls, v: Optional[float]) -> Optional[float]:
-        if v is None:
-            return v
-        if v < 0:
-            raise ValueError("Total spend amount must be non-negative")
-        return v
-
-    @field_validator("currency", mode="after")
-    @classmethod
-    def validate_currency(cls, v: Optional[str]) -> Optional[str]:
-        if v is None:
-            return v
-        normalized_v = v.strip().lower()
-        valid_currencies = [
-            "aud",
-            "cad",
-            "chf",
-            "cny",
-            "eur",
-            "gbp",
-            "jpy",
-            "sek",
-            "usd",
-        ]
-        if normalized_v not in valid_currencies:
-            raise ValueError(f"Currency must be one of: {valid_currencies}")
-        return normalized_v
-
-    @model_validator(mode="before")
-    @classmethod
-    def reject_null_institutional_code(cls, values: Any) -> Any:
-        # Key absent = "not updating" (PATCH semantics). An explicit null would
-        # silently clear the code, resolve no factor, and delete the entry's
-        # emissions. The payload mixin may carry the field top-level and/or
-        # under "data", so guard both shapes.
-        if not isinstance(values, dict):
-            return values
-        payloads = [values]
-        if isinstance(values.get("data"), dict):
-            payloads.append(values["data"])
-        for payload in payloads:
-            if (
-                "purchase_institutional_code" in payload
-                and payload["purchase_institutional_code"] is None
-            ):
-                raise ValueError("purchase_institutional_code cannot be null")
-        return values
-
-    @field_validator("purchase_institutional_code", mode="after")
-    @classmethod
-    def validate_purchase_institutional_code(cls, v: Optional[str]) -> Optional[str]:
-        # None here can only be the key-absent default (explicit null is
-        # rejected in the before-validator above); a blank/whitespace value
-        # provided on purpose must fail loudly here rather than silently
-        # resolving to no factor further down the pipeline.
-        if v is None:
-            return v
-        if not v.strip():
-            raise ValueError("purchase_institutional_code cannot be empty")
-        return v
-
-
-class PurchaseCentralizedHandlerUpdate(DataEntryUpdate):
-    name: Optional[str] = None
-    unit: Optional[str] = None
-    annual_consumption: Optional[float] = None
-    coef_to_kg: Optional[float] = None
-    note: Optional[str] = None
-
-    @field_validator("annual_consumption", "coef_to_kg", mode="after")
-    @classmethod
-    def validate_positive(cls, v: Optional[float]) -> Optional[float]:
-        if v is None:
-            return v
-        if v < 0:
-            raise ValueError("Value must be non-negative")
-        return v
-
-
-class PurchaseModuleHandler(BaseModuleHandler):
-    module_type: ModuleTypeEnum = ModuleTypeEnum.purchase
-    category_field: str = "purchase_category"
-    registration_keys = [
-        DataEntryTypeEnum.scientific_equipment,
-        DataEntryTypeEnum.it_equipment,
-        DataEntryTypeEnum.consumable_accessories,
-        DataEntryTypeEnum.biological_chemical_gaseous_product,
-        DataEntryTypeEnum.services,
-        DataEntryTypeEnum.vehicles,
-        DataEntryTypeEnum.other_purchases,
-    ]
-
-    create_dto = PurchaseHandlerCreate
-    update_dto = PurchaseHandlerUpdate
-    response_dto = PurchaseHandlerResponse
-
-    kind_field: str = "purchase_institutional_code"
-    # purchase_additional_code is optional on entries but is the primary
-    # factor key when present: it overrides the institutional-code match.
-    kind_field_override: Optional[str] = "purchase_additional_code"
-    subkind_field: Optional[str] = ""
-    # Required non-empty on create; update rejects present-but-blank/null
-    # (key-absent means "not updating"). CSV omits the key entirely when the
-    # cell is empty, so entries can still lack it — matching stays optional.
-    require_factor_to_match = False
-
-    sort_map = {
-        "id": DataEntry.id,
-        "name": DataEntry.data["name"].as_string(),
-        "supplier": DataEntry.data["supplier"].as_string(),
-        "quantity": DataEntry.data["quantity"].as_float(),
-        "total_spent_amount": DataEntry.data["total_spent_amount"].as_float(),
-        "currency": DataEntry.data["currency"].as_string(),
-        "purchase_institutional_code": DataEntry.data[
-            "purchase_institutional_code"
-        ].as_string(),
-        "kg_co2eq": DataEntryEmission.kg_co2eq,
-    }
-
-    filter_map = {
-        "name": DataEntry.data["name"].as_string(),
-        "supplier": DataEntry.data["supplier"].as_string(),
-        "purchase_institutional_code": DataEntry.data[
-            "purchase_institutional_code"
-        ].as_string(),
-    }
-
-    def to_response(
-        self,
-        data_entry: DataEntry,
-        enriched_data: dict | None = None,
-    ) -> PurchaseHandlerResponse:
-        data = enriched_data if enriched_data is not None else data_entry.data
-        return self.response_dto.model_validate(
-            {
-                "id": data_entry.id,
-                **data,
-                "data_entry_type_id": data_entry.data_entry_type_id,
-                "carbon_report_module_id": data_entry.carbon_report_module_id,
-                "name": data.get("name"),
-                "supplier": data.get("supplier"),
-                "quantity": data.get("quantity"),
-                "purchase_institutional_code": data.get("purchase_institutional_code"),
-                "total_spent_amount": data.get("total_spent_amount"),
-                "kg_co2eq": data.get("kg_co2eq", None),
-            }
-        )
-
-    def validate_create(self, payload: dict) -> PurchaseHandlerCreate:
-        return self.create_dto.model_validate(payload)
-
-    def validate_update(self, payload: dict) -> PurchaseHandlerUpdate:
-        return self.update_dto.model_validate(payload)
-
-    def resolve_computations(self, data_entry, emission_type, ctx: dict) -> list:
-
-        factor_id = ctx.get("primary_factor_id")
-        if factor_id is None:
-            return []
-
-        def _purchase_formula(ctx: dict, factor_values: dict) -> Optional[float]:
-            # Get the year to ensure we get the correct exchange rate for the year
-            # of the purchase
-            year = ctx.get("_year")
-            if year is None:
-                return None
-
-            total_spent_amount = ctx.get("total_spent_amount")
-            if total_spent_amount is None:
-                return None
-            entry_currency = (ctx.get("currency", "chf") or "chf").lower()
-            ef = factor_values.get("ef_kg_co2eq_per_currency")
-            if ef is None:
-                return None
-            ef_currency = (factor_values.get("currency", "eur") or "eur").lower()
-            if total_spent_amount is None or ef is None:
-                return None
-
-            # Use the exchange rate service to convert the total
-            # spent amount to the eur currency
-            total_spent_amount_eur = total_spent_amount
-            if entry_currency != "eur":
-                exchange_rate = ExchangeRatesService().get_exchange_rate_to_eur(
-                    year, entry_currency
-                )
-                total_spent_amount_eur = total_spent_amount * exchange_rate
-            # Similarly, convert the emission factor to eur if it's in a different
-            # currency, so that the final kg_co2eq is correctly computed in relation
-            # to the total spent amount in eur
-            ef_eur = ef
-            if ef_currency != "eur":
-                exchange_rate = ExchangeRatesService().get_exchange_rate_to_eur(
-                    year, ef_currency
-                )
-                ef_eur = ef * exchange_rate
-
-            return total_spent_amount_eur * ef_eur
-
-        return [
-            EmissionComputation(
-                emission_type=emission_type,
-                factor_id=factor_id,
-                formula_func=_purchase_formula,
-            )
-        ]
-
-
-class PurchaseCentralizedModuleHandler(BaseModuleHandler):
-    module_type: ModuleTypeEnum = ModuleTypeEnum.purchase
-    data_entry_type: DataEntryTypeEnum = DataEntryTypeEnum.purchases_centralized
-
-    create_dto = PurchaseCentralizedHandlerCreate
-    update_dto = PurchaseCentralizedHandlerUpdate
-    response_dto = PurchaseCentralizedHandlerResponse
-
-    kind_field: str = "name"
-    subkind_field: Optional[str] = ""
-
-    sort_map = {
-        "id": DataEntry.id,
-        "name": DataEntry.data["name"].as_string(),
-        "unit": DataEntry.data["unit"].as_string(),
-        "annual_consumption": DataEntry.data["annual_consumption"].as_float(),
-        "coef_to_kg": DataEntry.data["coef_to_kg"].as_float(),
-        "kg_co2eq": DataEntryEmission.kg_co2eq,
-    }
-
-    filter_map = {
-        "name": DataEntry.data["name"].as_string(),
-        "unit": DataEntry.data["unit"].as_string(),
-    }
-
-    def to_response(
-        self,
-        data_entry: DataEntry,
-        enriched_data: dict | None = None,
-    ) -> PurchaseCentralizedHandlerResponse:
-        data = enriched_data if enriched_data is not None else data_entry.data
-        return self.response_dto.model_validate(
-            {
-                "id": data_entry.id,
-                "data_entry_type_id": data_entry.data_entry_type_id,
-                "carbon_report_module_id": data_entry.carbon_report_module_id,
-                **data,
-            }
-        )
-
-    def validate_create(self, payload: dict) -> PurchaseCentralizedHandlerCreate:
-        return self.create_dto.model_validate(payload)
-
-    def validate_update(self, payload: dict) -> PurchaseCentralizedHandlerUpdate:
-        return self.update_dto.model_validate(payload)
-
-    def resolve_computations(self, data_entry, emission_type, ctx: dict) -> list:
-
-        factor_id = ctx.get("primary_factor_id")
-        if factor_id is None:
-            return []
-
-        def _additional_purchase_formula(ctx: dict, factor_values: dict):
-            annual_consumption = ctx.get("annual_consumption")
-            if annual_consumption is None:
-                return None
-            coef_to_kg = ctx.get("coef_to_kg")
-            if coef_to_kg is None:
-                return None
-            ef = factor_values.get("ef_kg_co2eq_per_kg")
-            if ef is None:
-                return None
-            return annual_consumption * coef_to_kg * ef
-
-        return [
-            EmissionComputation(
-                emission_type=emission_type,
-                factor_id=factor_id,
-                formula_func=_additional_purchase_formula,
-            )
-        ]
-
-
-## PURCHASE FACTOR HANDLERS
-
-# --- Additional Purchases ---
-
-purchase_additional_classification_fields: list[str] = ["name"]
-purchase_additional_value_fields: list[str] = ["ef_kg_co2eq_per_kg"]
-
-
-class PurchaseCentralizedFactorCreate(FactorCreate):
-    name: str
-    ef_kg_co2eq_per_kg: float
-
-    @field_validator("ef_kg_co2eq_per_kg", mode="after")
-    @classmethod
-    def validate_ef(cls, v: float) -> float:
-        if v < 0:
-            raise ValueError("ef_kg_co2eq_per_kg must be non-negative")
-        return v
-
-
-class PurchaseCentralizedFactorUpdate(FactorUpdate):
-    name: Optional[str] = None
-    ef_kg_co2eq_per_kg: Optional[float] = None
-
-    @field_validator("ef_kg_co2eq_per_kg", mode="after")
-    @classmethod
-    def validate_ef(cls, v: Optional[float]) -> Optional[float]:
-        if v is None:
-            return v
-        if v < 0:
-            raise ValueError("ef_kg_co2eq_per_kg must be non-negative")
-        return v
-
-
-class PurchaseCentralizedFactorResponse(FactorResponseGen):
-    name: str
-    ef_kg_co2eq_per_kg: float
-
-
-class PurchaseCentralizedFactorHandler(BaseFactorHandler):
-    data_entry_type: DataEntryTypeEnum | None = None
-    registration_keys = [DataEntryTypeEnum.purchases_centralized]
-    emission_type = EmissionType.purchases__centralized
-
-    create_dto = PurchaseCentralizedFactorCreate
-    update_dto = PurchaseCentralizedFactorUpdate
-    response_dto = PurchaseCentralizedFactorResponse
-
-    classification_fields: list[str] = purchase_additional_classification_fields
-    value_fields: list[str] = purchase_additional_value_fields
-
-
-# --- Common Purchases (7 types, same CSV format) ---
-
-purchase_common_classification_fields: list[str] = [
-    "purchase_institutional_code",
-    "purchase_additional_code",
-    "currency",
-]
-purchase_common_value_fields: list[str] = [
-    "ef_kg_co2eq_per_currency",
-    "translation_key",
-]
-
-
-class PurchaseCommonFactorCreate(FactorCreate):
-    currency: str
-    purchase_institutional_code: str
-    translation_key: Optional[str] = None
-    purchase_additional_code: Optional[str] = None
-    ef_kg_co2eq_per_currency: float
-    # purchase_category: str  # only for upload Mandatory (checked in csv upload)
-    # purchase_category is the routing column (picks the correct data_entry_type).
-    # It is consumed in the factor CSV provider, not carried on this DTO —
-    # its presence + case-sensitive enum is enforced
-    # there (see base_factor_csv_provider._resolve_data_entry_type).
-
-    @field_validator("ef_kg_co2eq_per_currency", mode="after")
-    @classmethod
-    def validate_ef(cls, v: float) -> float:
-        if v < 0:
-            raise ValueError("ef_kg_co2eq_per_currency must be non-negative")
-        return v
-
-    @field_validator("purchase_institutional_code", mode="after")
-    @classmethod
-    def validate_institutional_code(cls, v: str) -> str:
-        # Always present on factors: the additional-code-less rows are the
-        # per-institutional-code averages, so a factor without an
-        # institutional code could never be matched.
-        if not v.strip():
-            raise ValueError("purchase_institutional_code must not be empty")
-        return v
-
-
-class PurchaseCommonFactorUpdate(FactorUpdate):
-    purchase_institutional_code: Optional[str] = None
-    purchase_additional_code: Optional[str] = None
-    currency: Optional[str] = None
-    ef_kg_co2eq_per_currency: Optional[float] = None
-    translation_key: Optional[str] = None
-
-
-class PurchaseCommonFactorResponse(FactorResponseGen):
-    purchase_institutional_code: str
-    purchase_additional_code: Optional[str] = None
-    currency: str
-    ef_kg_co2eq_per_currency: Optional[float] = None
-    translation_key: Optional[str] = None
-
-
-class PurchaseCommonFactorHandler(BaseFactorHandler):
-    data_entry_type: DataEntryTypeEnum | None = None
-    registration_keys = [
-        DataEntryTypeEnum.scientific_equipment,
-        DataEntryTypeEnum.it_equipment,
-        DataEntryTypeEnum.consumable_accessories,
-        DataEntryTypeEnum.biological_chemical_gaseous_product,
-        DataEntryTypeEnum.services,
-        DataEntryTypeEnum.vehicles,
-        DataEntryTypeEnum.other_purchases,
-    ]
-    category_field: str = "purchase_category"
-    emission_type = None  # resolved per type via DATA_ENTRY_TO_EMISSION_TYPES
-
-    create_dto = PurchaseCommonFactorCreate
-    update_dto = PurchaseCommonFactorUpdate
-    response_dto = PurchaseCommonFactorResponse
-
-    classification_fields: list[str] = purchase_common_classification_fields
-    value_fields: list[str] = purchase_common_value_fields
