+++ backend/app/seed/seed_generic_factors.py
@@ -4,21 +4,13 @@ from pathlib import Path
 
 from sqlmodel.ext.asyncio.session import AsyncSession
 
+import app.modules  # noqa: F401  # registers all module + factor handlers
 from app.core.config import get_settings
 from app.core.logging import get_logger
 from app.db import SessionLocal
 from app.models.data_entry import DataEntryTypeEnum
 from app.models.data_ingestion import TargetType
 from app.models.module_type import get_module_type_for_data_entry_type
-from app.modules.external_cloud_and_ai import schemas as schemas
-from app.modules.process_emissions import schemas as _pe_schemas  # noqa: F401
-from app.modules.purchase import schemas as _purchase_schemas  # noqa: F401
-from app.modules.research_facilities import (
-    animals_schemas as _rf_animals_schemas,  # noqa: F401
-)
-from app.modules.research_facilities import (
-    common_schemas as _rf_common_schemas,  # noqa: F401
-)
 from app.schemas.factor import BaseFactorHandler
 from app.seed._stub_jobs import create_seed_stub_job
 from app.services.data_ingestion.csv_providers.local_seed import LocalFactorCSVProvider
